requirements.txt
asgiref==3.7.2
certifi==2023.7.22
charset-normalizer==3.2.0
crossrefapi==1.6.0
Django==4.2.4
django-cors-headers==4.3.0
django-environ==0.10.0
django-js-asset==2.1.0
djangorestframework==3.14.0
gunicorn==21.2.0
idna==3.4
packaging==23.1
psycopg2-binary==2.9.9
python-dotenv==1.0.0
pytz==2023.3
requests==2.31.0
sqlparse==0.4.4
typing_extensions==4.7.1
urllib3==1.26.16
